remove old bunded test and add k8s bits to otel native bundled test
fmt

fmt

config tweak and bundle
