+++ tests/receivers/postgresql/testdata/resource_metrics/all_k8s.yaml
@@ -1,0 +1,125 @@
+resource_metrics:
+  - attributes:
+      postgresql.database.name: <ANY>
+      k8s.namespace.name: <ANY>
+      k8s.pod.name: <ANY>
+      k8s.pod.uid: <ANY>
+    scope_metrics:
+      - instrumentation_scope:
+          name: otelcol/postgresqlreceiver
+          version: <ANY>
+          attributes: {}
+        metrics:
+          - name: postgresql.commits
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.db_size
+            type: IntNonmonotonicCumulativeSum
+          - name: postgresql.rollbacks
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.table.count
+            type: IntNonmonotonicCumulativeSum
+  - attributes:
+      postgresql.database.name: <ANY>
+      postgresql.index.name: <ANY>
+      postgresql.table.name: <ANY>
+      k8s.namespace.name: <ANY>
+      k8s.pod.name: <ANY>
+      k8s.pod.uid: <ANY>
+    scope_metrics:
+      - instrumentation_scope:
+          name: otelcol/postgresqlreceiver
+          version: <ANY>
+          attributes: {}
+        metrics:
+          - name: postgresql.index.scans
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.index.size
+            type: IntGauge
+  - attributes:
+      postgresql.database.name: <ANY>
+      postgresql.table.name: <ANY>
+      k8s.namespace.name: <ANY>
+      k8s.pod.name: <ANY>
+      k8s.pod.uid: <ANY>
+    scope_metrics:
+      - instrumentation_scope:
+          name: otelcol/postgresqlreceiver
+          version: <ANY>
+          attributes: {}
+        metrics:
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: heap_read
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: heap_hit
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: idx_read
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: idx_hit
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: toast_hit
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: tidx_read
+          - name: postgresql.blocks_read
+            type: IntMonotonicCumulativeSum
+            attributes:
+              source: tidx_hit
+          - name: postgresql.operations
+            type: IntMonotonicCumulativeSum
+            attributes:
+              operation: del
+          - name: postgresql.operations
+            type: IntMonotonicCumulativeSum
+            attributes:
+              operation: upd
+          - name: postgresql.operations
+            type: IntMonotonicCumulativeSum
+            attributes:
+              operation: hot_upd
+          - name: postgresql.rows
+            type: IntNonmonotonicCumulativeSum
+            attributes:
+              state: dead
+          - name: postgresql.rows
+            type: IntNonmonotonicCumulativeSum
+            attributes:
+              state: live
+          - name: postgresql.table.size
+            type: IntNonmonotonicCumulativeSum
+          - name: postgresql.table.vacuum.count
+            type: IntMonotonicCumulativeSum
+  - attributes: 
+      k8s.namespace.name: <ANY>
+      k8s.pod.name: <ANY>
+      k8s.pod.uid: <ANY>
+    scope_metrics:
+      - instrumentation_scope:
+          name: otelcol/postgresqlreceiver
+          version: <ANY>
+          attributes: {}
+        metrics:
+          - name: postgresql.bgwriter.buffers.allocated
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.bgwriter.buffers.writes
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.bgwriter.checkpoint.count
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.bgwriter.duration
+            type: DoubleMonotonicCumulativeSum
+          - name: postgresql.bgwriter.maxwritten
+            type: IntMonotonicCumulativeSum
+          - name: postgresql.connection.max
+            type: IntGauge
+          - name: postgresql.database.count
+            type: IntNonmonotonicCumulativeSum
